+++ R/agree_nest.R
@@ -1,30 +1,37 @@
 #' Tests for Absolute Agreement with Nested Data
-#' @description agree_nest produces an absolute agreement analysis for data where there is multiple observations per subject but the mean varies within subjects as described by Zou (2013). Output mirrors that of agree_test but CCC is calculated via U-statistics.
+#'
+#' @description
+#'
+#' `r lifecycle::badge('superseded')`
+#'
+#' Development on `agree_nest()` is complete, and for new code we recommend
+#' switching to `agreement_limit()`, which is easier to use, has more features,
+#' and still under active development.
+#'
+#' `agree_nest` produces an absolute agreement analysis for data where there is multiple observations per subject but the mean varies within subjects as described by Zou (2013). Output mirrors that of agree_test but CCC is calculated via U-statistics.
 #' @param x Name of column with first measurement
 #' @param y Name of other column with the other measurement to compare to the first.
 #' @param id Column with subject identifier
 #' @param data Data frame with all data
-#' @param conf.level the confidence level required. Default is 95\%.
-#' @param agree.level the agreement level required. Default is 95\%. The proportion of data that should lie between the thresholds, for 95\% limits of agreement this should be 0.95.
+#' @param conf.level the confidence level required. Default is 95%.
+#' @param agree.level the agreement level required. Default is 95%. The proportion of data that should lie between the thresholds, for 95% limits of agreement this should be 0.95.
 #' @param delta The threshold below which methods agree/can be considered equivalent, can be in any units. Equivalence Bound for Agreement.
 #' @param prop_bias Logical indicator (TRUE/FALSE) of whether proportional bias should be considered for the limits of agreement calculations.
 #' @param TOST Logical indicator (TRUE/FALSE) of whether to use two one-tailed tests for the limits of agreement. Default is TRUE.
-#'
+#' @param ccc Calculate concordance correlation coefficient.
 #' @return Returns single simple_agree class object with the results of the agreement analysis.
 #'
-#' \describe{
-#'   \item{\code{"loa"}}{A data frame of the limits of agreement including the average difference between the two sets of measurements, the standard deviation of the difference between the two sets of measurements and the lower and upper confidence limits of the difference between the two sets of measurements.}
-#'   \item{\code{"h0_test"}}{Decision from hypothesis test.}
-#'   \item{\code{"ccc.xy"}}{Lin's concordance correlation coefficient and confidence intervals using U-statistics. Warning: if underlying value varies this estimate will be inaccurate.}
-#'   \item{\code{"call"}}{the matched call}
-#'   \item{\code{"var_comp"}}{Table of Variance Components}
-#'   \item{\code{"class"}}{The type of simple_agree analysis}
-#' }
-
+#'   - `loa`: A data frame of the limits of agreement including the average difference between the two sets of measurements, the standard deviation of the difference between the two sets of measurements and the lower and upper confidence limits of the difference between the two sets of measurements.
+#'   - `h0_test`: Decision from hypothesis test.
+#'   - `ccc.xy`: Lin's concordance correlation coefficient and confidence intervals using U-statistics. Warning: if underlying value varies this estimate will be inaccurate.
+#'   - `call`: the matched call.
+#'   - `var_comp`: Table of Variance Components.
+#'   - `class`: The type of simple_agree analysis.
+#'
 #' @examples
 #' data('reps')
 #' agree_nest(x = "x", y = "y", id = "id", data = reps, delta = 2)
-#' @section References:
+#' @references
 #' Zou, G. Y. (2013). Confidence interval estimation for the Bland–Altman limits of agreement with multiple observations per individual. Statistical methods in medical research, 22(6), 630-642.
 #'
 #' King, TS and Chinchilli, VM. (2001). A generalized concordance correlation coefficient for continuous and categorical data. Statistics in Medicine, 20, 2131:2147.
@@ -47,8 +54,9 @@ agree_nest <- function(x,
                        agree.level = .95,
                        conf.level = .95,
                        TOST = TRUE,
-                       prop_bias = FALSE){
-
+                       prop_bias = FALSE,
+                       ccc = TRUE){
+  lifecycle::deprecate_soft("0.2.0", "agree_nest()", "agreement_limit()")
   agreeq = qnorm(1 - (1 - agree.level) / 2)
   agree_l = 1 - (1 - agree.level) / 2
   agree_u = (1 - agree.level) / 2
@@ -79,6 +87,7 @@ agree_nest <- function(x,
                  names_to = "method",
                  values_to = "measure")
 
+  if(ccc){
   ccc_nest = cccUst(dataset = df_long,
                     ry = "measure",
                     rmet = "method",
@@ -88,6 +97,12 @@ agree_nest <- function(x,
                       lower.ci = ccc_nest[2],
                       upper.ci = ccc_nest[3],
                       SE = ccc_nest[4])
+  } else{
+    ccc.xy = data.frame(est.ccc = NA,
+                        lower.ci = NA,
+                        upper.ci = NA,
+                        SE = NA)
+  }
   df_lmer = df %>%
     mutate(mean = (x+y)/2,
            delta = x - y)